# std.uni.isNumber: significant performance increase for ASCII