When testing the old version vs this version on `dchar a = '7';`, I got the following results for 1 million calls:

```
isNumber		85 ms, 914 μs, and 3 hnsecs
isNumber2		4 ms and 368 μs
```