@Hackerpilot 2016-04-01
Auto-merge toggled on