@deadalnix 2016-04-01
I second @andralex . magic constants in the code makes it inscrutable and fragile very quickly.