@deadalnix 2016-04-01
Also, if that is possible, it'd be great to find at compile time what is the lowest value a number can have outside of the ASCII range and use that instead of 0xAA .