@Hackerpilot 2016-04-01
Ah. When I saw your "fixed" comment I didn't actually check that you replaced the constants.